Splits SassTask tests to unit and integration ones
